chore: audit follow-ups — CLAUDE.md, shared utils, test fixtures, size budgets
Bundles four small but useful internal cleanups identified in the post-audit
review. No public API or runtime behavior changes.

CLAUDE.md
  Refreshed with everything learned from the audit + refactor PRs:
  - Lerna -> Changesets (PR #147), updated single-package build cmd
  - Package count corrected (15, was claiming 16)
  - Document where rocketstyle chain methods come from (Object.assign vs
    createStaticsChainingEnhancers via STATIC_KEYS) — prevents future
    accidental shadowing like the one we caught in PR #170
  - Note the Overlay positionMath split as the reference pattern for
    splitting fat hooks (PR #168)
  - Type system notes: chain machinery TS limitation, internal `any` is OK
    when interface is properly typed (PR #170), useExhaustiveDependencies
    pattern for stable factory-closure-captured options
  - Performance hot paths section listing the PR #166 wins so future edits
    don't regress them
  - New `bun run size` command listed under dev commands

styler: shared evictMapByPercent utility (PR #166 follow-up)
  `sheet.ts` and `resolve.ts` had two independent copies of the same
  "evict oldest 10% by insertion order" loop. Extracted to a new
  `packages/styler/src/evict.ts` (separate file to avoid the
  shared.ts <-> resolve.ts circular dep).

kinetic: shared test fixtures
  6 test files (Transition, TransitionGroup, Stagger, Collapse,
  TransitionItem, kinetic) each reimplemented the same matchMedia mock,
  rAF stub, and `fireTransitionEnd` helper — ~50 lines each, ~300 lines
  total. Extracted to `packages/kinetic/src/__tests__/setupFixtures.ts`:
    - `setupMatchMedia()` — installs the matchMedia mock
    - `setupRaf()` — controllable rAF queue + fake timers, returns flushRaf
    - `fireTransitionEnd(el)` / `fireAnimationEnd(el)` — jsdom event helpers
  Net diff: -300 lines from tests + 80 lines fixtures.

Bundle size budgets
  New `.size-limit.json` config + `bun run size` script + CI step. Budgets
  set ~20% above current sizes to allow normal growth while catching real
  regressions. Uses `@size-limit/file` (just measures the built bundle —
  doesn't try to bundle peer deps like react-native).

  Current sizes (gzipped) vs budget:
    @vitus-labs/connector-styled-components: 186 B / 300 B
    @vitus-labs/connector-styler:            193 B / 300 B
    @vitus-labs/connector-emotion:           1.10 KB / 1.4 KB
    @vitus-labs/attrs:                       2.52 KB / 3 KB
    @vitus-labs/connector-native:            2.74 KB / 3.5 KB
    @vitus-labs/coolgrid:                    3.52 KB / 4.5 KB
    @vitus-labs/core:                        4.83 KB / 6 KB
    @vitus-labs/kinetic:                     5.09 KB / 6.5 KB
    @vitus-labs/hooks:                       5.25 KB / 6.5 KB
    @vitus-labs/kinetic-presets:             6.32 KB / 8 KB
    @vitus-labs/rocketstyle:                 7.18 KB / 9 KB
    @vitus-labs/rocketstories:               8.27 KB / 10 KB
    @vitus-labs/unistyle:                    9.27 KB / 12 KB
    @vitus-labs/styler:                      9.73 KB / 12 KB
    @vitus-labs/elements:                    11.82 KB / 15 KB
    Total: ~78 KB gzipped across 15 packages

All 2599 tests pass. 0 type errors. 0 lint errors. 15 packages build.

Skipped (from audit candidates):
- ReactElement<any> -> ReactElement: would propagate as ReactElement<unknown>
  in newer @types/react and break props access in Transition impl files.
  Cascade not worth a marginal idiom change.
- styler hotCache + staticComponentCache consolidation: code is dense but
  correct, refactoring risks breaking subtle ordering for marginal gain.
